Changes in structure IShippingRateCalculationProvider plugins, move away from viewcomponent, now use Controllers
