+++ src/Plugins/Shipping.ByWeight/ByWeightShippingProvider.cs
@@ -312,7 +312,7 @@ namespace Shipping.ByWeight
             return await Task.FromResult(new List<string>());
         }
 
-        public async Task<string> GetPublicViewComponentName()
+        public async Task<string> GetControllerRouteName()
         {
             return await Task.FromResult("");
         }
