+++ src/Plugins/Shipping.ShippingPoint/Views/SelectedShippingPoint/Points.cshtml
@@ -1,4 +1,7 @@
 ﻿@model List<SelectListItem>
+@{
+    Layout = "";
+}
 <div class="script-tag">
     <script>
         function renderShippingOptionInfo(param) {
